pyproject.toml
[tool.poetry]
name = "ec-pov-monitor"
version = "0.1.0"
description = "Ecuador Economy & Poverty Monitor (official published series only)"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "ec_poverty_monitor", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
httpx = "^0.27.0"
pandas = "^2.2.2"
duckdb = "^1.0.0"
PyYAML = "^6.0.1"
click = "8.1.7"
typer = {version = "^0.12.3", extras = ["all"]}
rich = "^13.7.1"
python-dateutil = "^2.9.0.post0"
streamlit = "^1.36.0"
altair = "^5.3.0"
lxml = "^5.3.0"
openpyxl = "^3.1.5"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
ruff = "^0.5.5"

[tool.poetry.scripts]
ecmon = "ec_poverty_monitor.cli:app"

[tool.ruff]
line-length = 100
target-version = "py311"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
